Add files via upload
Had made quite a lot of changes compare to previous versions of my code, as a result the css part is a bit messy as some of the class might not exist anymore (combined with other parts). The html part should be easy to understand though. Have used some libraries which have mentioned in previous meetings (just for us to reference, may need to remove and create some functions on our own before submission.